Separate user SDK and internal developer documentation
