+++ jsdoc/config-sdk.json
@@ -1,0 +1,24 @@
+{
+    "opts": {
+        "template": "hifi-jsdoc-template"
+    },
+    "docdash": {
+        "meta": {                       
+            "title": "",                
+            "description": "",          
+            "keyword": ""               
+        },
+        "search": [true],  
+        "collapse": [true],       
+        "typedefs": [false]    
+    },    
+    "templates": {
+        "default": {
+            "outputSourceFiles": false
+        }
+    },
+    "plugins": [
+        "plugins/vircadia",
+        "plugins/vircadia-sdk"
+    ]
+}
