+++ jsdoc/dev-mainpage.md
@@ -1,0 +1,16 @@
+<h2 class="subheading">Developer Documention</h2>
+
+The <strong>Vircadia Web SDK</strong> is a JavaScript SDK for developing web-based clients for virtual worlds powered by
+Vircadia. Vircadia domain servers provide the worlds (a.k.a. "domains") to visit, and the Vircadia metaverse server provides
+global services that connect the users and domains.
+See the user docs to [Understand the Architecture](https://docs.vircadia.dev/explore/get-started/architecture.html).
+
+This SDK provides interfaces to:
+- Connect to domains.
+- Use metaverse services.
+
+<hr />
+
+For scripting API documentation, see the [Vircadia API Reference](https://apidocs.vircadia.dev).
+
+To learn more about using Vircadia and exploring the metaverse, see the [User Documentation](https://docs.vircadia.dev).
