+++ jsdoc/plugins/vircadia.js
@@ -1,54 +1,4 @@
-function endsWith(path, exts) {
-    var result = false;
-    exts.forEach(function(ext) {
-        if (path.endsWith(ext)) {
-            result = true;
-        }
-    });
-    return result;
-}
-
 exports.handlers = {
-
-    // This event is triggered before parsing has even started.
-    // We use this event to scan the C++ files for jsdoc comments
-    // and reformat them into a form digestable by jsdoc.
-    beforeParse: function(e) {
-        var pathTools = require('path');
-        var rootFolder = pathTools.dirname(e.filename);
-        console.log("Scanning the Vircadia source for JSDoc comments...");
-
-        // directories to scan for jsdoc comments
-        var dirList = [
-            '.'
-        ];
-
-        // only files with this extension will be searched for jsdoc comments.
-        var exts = ['.h', '.cpp'];
-
-        var fs = require('fs');
-        dirList.forEach(function (dir) {
-            var joinedDir = pathTools.join(rootFolder, dir);
-            var files = fs.readdirSync(joinedDir);
-            files.forEach(function (file) {
-                var path = pathTools.join(joinedDir, file);
-                if (fs.lstatSync(path).isFile() && endsWith(path, exts)) {
-                    // load entire file into a string
-                    var data = fs.readFileSync(path, "utf8");
-
-                    // this regex searches for blocks starting with /**jsdoc and end with */
-                    var reg = /(\/\*\*jsdoc(.|[\r\n])*?\*\/)/gm;
-                    var matches = data.match(reg);
-                    if (matches) {
-                        // add to source, but strip off c-comment asterisks
-                        e.source += matches.map(function (s) {
-                            return s.replace('/**jsdoc', '/**');
-                        }).join('\n');
-                    }
-                }
-            });
-        });
-    },
 
     // This event is triggered when a new doclet has been created
     // but before it is passed to the template for output
