+++ src/Vircadia.js
@@ -10,9 +10,10 @@
 //  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
 //
 
-/**
+/*@sdkdoc
  * This is the Vircadia SDK.
  * @namespace Vircadia
+ * @sdk
  */
 const Vircadia = (function () {
